removed old version of quiz and updated router
