--- application/src/main/resources/static/js/react-client/src/pages/quiz/newIndex.js
@@ -1,110 +1,0 @@
-import React, { Component } from 'react';
-import Quiz from "../../components/quiz/quiz";
-import quizQuestions from "../../components/quiz/quizQuestions";
-import Result from "../../components/quiz/result";
-import style from "./style.module.scss";
-
-class quizPage extends Component {
-  constructor(props) {
-    super(props);
-
-    this.state = {
-      counter: 0,
-      questionId: 1,
-      question: '',
-      answerOptions: [],
-      answer: '',
-      answersCount: {},
-      result: ''
-    };
-
-    this.handleAnswerSelected = this.handleAnswerSelected.bind(this);
-  }
-
-  componentDidMount() {
-    this.setState({
-      question: quizQuestions[0].question,
-      answerOptions: quizQuestions[0].answers
-    });
-  }
-
-  setUserAnswer(answer) {
-    this.setState((state) => ({
-      answersCount: {
-        ...state.answersCount,
-        [answer]: (state.answersCount[answer] || 0) + 1
-      },
-      answer: answer
-    }));
-  }
-
-  handleAnswerSelected(e) {
-    this.setUserAnswer(e.currentTarget.value);
-    if (this.state.questionId < quizQuestions.length) {
-      setTimeout(() => this.setNextQuestion(), 300);
-    } else {
-      setTimeout(() => this.setResults(this.getResults()), 300);
-    }
-  }
-
-  setNextQuestion() {
-    const counter = this.state.counter + 1;
-    const questionId = this.state.questionId + 1;
-    this.setState({
-      counter: counter,
-      questionId: questionId,
-      question: quizQuestions[counter].question,
-      answerOptions: quizQuestions[counter].answers,
-      answer: ''
-    });
-  }
-
-  getResults() {
-    const answersCount = this.state.answersCount;
-    const answersCountKeys = Object.keys(answersCount);
-    const answersCountValues = answersCountKeys.map((key) => answersCount[key]);
-    const maxAnswerCount = Math.max.apply(null, answersCountValues);
-
-    return answersCountKeys.filter((key) => answersCount[key] === maxAnswerCount);
-  }
-
-  setResults (result) {
-    if (result.length === 1) {
-      this.setState({ result: result[0] });
-    } else {
-      this.setState({ result: 'Undetermined' });
-    }
-  }
-
-  renderQuiz() {
-    return (
-      <Quiz
-        answer={this.state.answer}
-        answerOptions={this.state.answerOptions}
-        questionId={this.state.questionId}
-        question={this.state.question}
-        questionTotal={quizQuestions.length}
-        onAnswerSelected={this.handleAnswerSelected}
-      />
-    );
-  }
-  
-  renderResult() {
-    return (
-      <Result quizResult={this.state.result} />
-    );
-  }
-
-  render() {
-    return (
-      <div className={style.quiz}>
-        <div className={style.header}>
-          <h2>Foot Ailment Quiz</h2>
-        </div>
-        {this.state.result ? this.renderResult() : this.renderQuiz()}
-      </div>
-    );
-  }
-}
-
-export default quizPage;
